Set `skipdist = True`.
Running the unittests directly from the project directory currently takes
0.11s.  However, running the same through tox takes 5+ seconds.

Debugging points to the `pip install` it runs internally.

Turning this off bring it back down to about 0.9s to run coverage and report,
which feels like a good compromise.

Obviously, if this project gets bigger, then the time might be dominated by
runtime instead of setup time.  But for now, skip it.